Extract classes out of DeadVarElim
